J.A.R.V.I.S/requirements.txt
pygame==2.6.1
google-cloud-texttospeech==2.18.0
openai==0.28.0
requests==2.32.3
phue==1.1
SpeechRecognition==3.10.4
